requirements_dev.txt
-r requirements.txt
asgiref==3.5.2
atomicwrites==1.4.1
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.0.12
colorama==0.4.5
Django==2.2.16
django-extensions==2.2.9
django-filter==2.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.1
flake8==5.0.4
idna==3.4
importlib-metadata==4.2.0
iniconfig==1.1.1
isort==5.10.1
mccabe==0.7.0
numpy==1.21.6
packaging==21.3
pandas==1.3.5
plotly==5.10.0
pluggy==0.13.1
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.1.0
pyparsing==3.0.9
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
pytz==2022.4
requests==2.26.0
six==1.16.0
sqlparse==0.4.3
tenacity==8.1.0
toml==0.10.2
typing-extensions==4.4.0
urllib3==1.26.12
zipp==3.9.0